Implement IntoTraversal for Vec<T>
